+++ test/firing-state.test.ts
@@ -1,16 +1,30 @@
 import { range } from 'lodash';
 import * as mat from 'transformation-matrix';
 
-import { FiringState, IFireData, IFireDataModifier } from 'guntree/firing-state';
+import { FiringState, IFireData, IFireDataModifier, IRepeatStateManager } from 'guntree/firing-state';
 import { IPlayer } from 'guntree/player';
 import { IBullet } from 'guntree/gun';
 import { IMuzzle } from 'guntree/muzzle';
 import { simpleMock } from './util';
 
+const createSimpleFireDataMock = () => {
+    const fireData = simpleMock<IFireData>();
+    const fireDataClone = simpleMock<IFireData>();
+    fireData.copy = jest.fn().mockReturnValueOnce(fireDataClone);
+    return [fireData, fireDataClone];
+};
+
+const createSimpleRSMMock = () => {
+    const rsm = simpleMock<IRepeatStateManager>();
+    const rsmClone = simpleMock<IRepeatStateManager>();
+    rsm.copy = jest.fn().mockReturnValueOnce(rsmClone);
+    return [rsm, rsmClone];
+};
+
 describe('#FiringState', () => {
     test('can add modifier and apply their', () => {
-        // Given FireData as clone
-        const fireDataClone = simpleMock<IFireData>();
+        // Given FireData with clone
+        const [fireData, fireDataClone] = createSimpleFireDataMock();
         fireDataClone.transform = mat.translate(0);
 
         // And muzzle with transform
@@ -18,7 +32,9 @@ describe('#FiringState', () => {
         muzzle.getMuzzleTransform = jest.fn().mockReturnValueOnce(mat.translate(0));
 
         // And FiringState with muzzle and FireData clone
-        const state = new FiringState(simpleMock<IPlayer>());
+        const state = new FiringState(
+            simpleMock<IPlayer>(), fireData, simpleMock<IRepeatStateManager>(),
+        );
         state.muzzle = muzzle;
         state.fireData.copy = jest.fn().mockReturnValueOnce(fireDataClone);
 
@@ -47,39 +63,51 @@ describe('#FiringState', () => {
     });
 
     test('can copy with fireData', () => {
-        // Given FiringState
-        const state = new FiringState(simpleMock<IPlayer>());
+        // Given FireData with clone
+        const [fireData, fireDataClone] = createSimpleFireDataMock();
 
-        // And FiringState's fireData's copy was pre-defined
-        const dataClone = jest.fn();
-        state.fireData.copy = jest.fn().mockReturnValueOnce(dataClone);
+        // And RepeatStateManager with clone
+        const [rsm, rsmClone] = createSimpleRSMMock();
+
+        // And FiringState
+        const state = new FiringState(simpleMock(), fireData, rsm);
 
         // When copy FiringState
         const clone = state.copy();
 
         // Then copy's fireData is original's clone
-        expect(clone.fireData).toBe(dataClone);
+        expect(clone.fireData).toBe(fireDataClone);
     });
 
     test('can copy with repeatStates', () => {
-        // Given FiringState
-        const state = new FiringState(simpleMock<IPlayer>());
+        // Given FireData with clone
+        const [fireData, fireDataClone] = createSimpleFireDataMock();
 
-        // And FiringState's repeatStates's copy was pre-defined
-        const rsClone = jest.fn();
-        state.repeatStates.copy = jest.fn().mockReturnValueOnce(rsClone);
+        // And RepeatStateManager with clone
+        const [rsm, rsmClone] = createSimpleRSMMock();
+
+        // And FiringState
+        const state = new FiringState(simpleMock(), fireData, rsm);
 
         // When copy FiringState
         const clone = state.copy();
 
         // Then copy's repeatStates is original's clone
-        expect(clone.repeatStates).toBe(rsClone);
+        expect(clone.repeatStates).toBe(rsmClone);
     });
 
     test('can copy with muzzle', () => {
-        // Given FiringState with muzzle
-        const state = new FiringState(simpleMock<IPlayer>());
+        // Given FireData with clone
+        const [fireData, fireDataClone] = createSimpleFireDataMock();
+
+        // And RepeatStateManager with clone
+        const [rsm, rsmClone] = createSimpleRSMMock();
+
+        // And Muzzle
         const muzzle = simpleMock<IMuzzle>();
+
+        // Given FiringState with muzzle
+        const state = new FiringState(simpleMock(), fireData, rsm);
         state.muzzle = muzzle;
 
         // When copy FiringState
@@ -96,32 +124,35 @@ describe('#FiringState', () => {
         player.getMuzzle = jest.fn().mockReturnValueOnce(muzzle);
 
         // And FiringState
-        const state = new FiringState(player);
+        const state = new FiringState(player, simpleMock(), simpleMock());
 
         // When get muzzle by name
         const muzzleName = 'a';
-        const currentMuzzle = state.getMuzzleByName(muzzleName);
+        const gottenMuzzle = state.getMuzzleByName(muzzleName);
 
-        // Then gotten muzzle is Player's muzzle from player
-        expect(currentMuzzle).toBe(muzzle);
+        // Then gotten muzzle is Player's muzzle
+        expect(gottenMuzzle).toBe(muzzle);
         expect(player.getMuzzle).toBeCalledWith(muzzleName);
     });
 
     test('can call fire and pass firing to muzzle', () => {
-        // Given FiringState with muzzle.
+        // Given Muzzle can fire
         const muzzle = simpleMock<IMuzzle>();
         muzzle.fire = jest.fn();
-        const modifiedFireData = simpleMock<IFireData>();
-        const state = new FiringState(simpleMock<IPlayer>());
+
+        // And RepeatStateManager with clone
+        const rsm = simpleMock<IRepeatStateManager>();
+        const rsmClone = simpleMock<IRepeatStateManager>();
+        rsm.copy = jest.fn().mockReturnValueOnce(rsmClone);
+
+        // And FiringState with muzzle and can calcModifiedFireData
+        const state = new FiringState(simpleMock(), simpleMock(), rsm);
         state.muzzle = muzzle;
+        const modifiedFireData = simpleMock<IFireData>();
         state.calcModifiedFireData = jest.fn().mockReturnValueOnce(modifiedFireData);
 
         // And Bullet
         const bullet = simpleMock<IBullet>();
-
-        // And FiringState's repeatStates's copy was pre-defined
-        const rsClone = jest.fn();
-        state.repeatStates.copy = jest.fn().mockReturnValueOnce(rsClone);
 
         // When fire
         state.fire(bullet);
@@ -131,13 +162,20 @@ describe('#FiringState', () => {
     });
 
     test('use muzzle transform in calcModifiedFireData', () => {
-        // Given muzzle
+        // Given FireData with clone
+        const [fireData, fireDataClone] = createSimpleFireDataMock();
+        fireDataClone.transform = mat.translate(0);
+
+        // And RepeatStateManager with clone
+        const [rsm, rsmClone] = createSimpleRSMMock();
+
+        // And muzzle
         const muzzle = simpleMock<IMuzzle>();
         const muzzleTrans = mat.translate(1.2, 5.3);
         muzzle.getMuzzleTransform = jest.fn().mockReturnValueOnce(muzzleTrans);
 
         // And firing state with muzzle and FireData clone
-        const state = new FiringState(simpleMock<IPlayer>());
+        const state = new FiringState(simpleMock(), fireData, rsm);
         state.muzzle = muzzle;
 
         // When calc modified fireData