+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [0.4.3] - 2019-04-25
+### Changed
+- Update FiringState to used dependency injection
+- Update Player to used dependency injection
+- Delete initial parameters and texts from Player.
+- Add initial parameters to FireData.
+
 ## [0.4.2] - 2019-04-25
 ### Fixed
 - Fix .gitlab-ci.yml